+++ lib/src/celestialBodies/sun.dart
@@ -3,7 +3,7 @@ import 'dart:math';
 import 'package:solar_calculator/solar_calculator.dart';
 import 'package:solar_calculator/src/celestialBodies/earth.dart';
 import 'package:solar_calculator/src/coordinateSystems/horizontalCoordinate.dart';
-import 'package:solar_calculator/src/julianDate.dart';
+import 'package:solar_calculator/src/instant.dart';
 import 'package:solar_calculator/src/math.dart';
 import 'package:solar_calculator/src/extensions.dart';
 
@@ -11,7 +11,7 @@ class Sun {
   // Cache
   static final Map<double, Sun> _cache = <double, Sun>{};
 
-  final JulianDate julianDate;
+  final Instant instant;
 
   double? _meanLongitude;
   double? _meanAnomaly;
@@ -25,10 +25,10 @@ class Sun {
 
   EquatorialCoordinate? _equatorialPosition;
 
-  factory Sun(JulianDate julianDate) =>
-      _cache.putIfAbsent(julianDate.julianDay, () => Sun._internal(julianDate));
+  factory Sun(Instant instant) =>
+      _cache.putIfAbsent(instant.julianDay, () => Sun._internal(instant));
 
-  Sun._internal(this.julianDate);
+  Sun._internal(this.instant);
 
   /// The equation of time in minutes.
   ///
@@ -45,16 +45,16 @@ class Sun {
   /// Earth's elliptical orbit and Kepler's law of equal areas in equal times are the culprits behind this phenomenon.
   double get equationOfTime {
     if (_equationOfTime == null) {
-      var earth = Earth(julianDate);
+      final earth = Earth(instant);
 
-      var epsilonRadians = earth.correctedObliquityOfEcliptic.toRadians();
-      var l0Radians = meanLongitude.toRadians();
-      var e = earth.orbitalEccentricity;
-      var mRadians = uncorrectedMeanAnomaly.toRadians();
+      final epsilonRadians = earth.correctedObliquityOfEcliptic.toRadians();
+      final l0Radians = meanLongitude.toRadians();
+      final e = earth.orbitalEccentricity;
+      final mRadians = uncorrectedMeanAnomaly.toRadians();
 
-      var y = pow(tan(epsilonRadians / 2), 2);
+      final y = pow(tan(epsilonRadians / 2), 2);
 
-      var eTime = (y * sin(2.0 * l0Radians)) -
+      final eTime = (y * sin(2.0 * l0Radians)) -
           (2.0 * e * sin(mRadians)) +
           (4.0 * e * y * sin(mRadians) * cos(2.0 * l0Radians)) -
           (0.5 * y * y * sin(4.0 * l0Radians)) -
@@ -70,15 +70,9 @@ class Sun {
   ///
   /// The mean longitude is the ecliptic longitude at which an orbiting body could be found if its orbit were circular and
   /// free of perturbations.
-  double get meanLongitude {
-    if (_meanLongitude == null) {
-      var l0 = evaluatePolynomial(
-          julianDate.julianCenturies, [280.46646, 36000.76983, 0.0003032]);
-      _meanLongitude = l0.correctDegreesForLargeAngles();
-    }
-
-    return _meanLongitude!;
-  }
+  double get meanLongitude => _meanLongitude ??= evaluatePolynomial(
+          instant.julianCenturies, [280.46646, 36000.76983, 0.0003032])
+      .correctDegreesForLargeAngles();
 
   /// The geometric mean anomaly of the Sun in degrees.
   ///
@@ -97,7 +91,7 @@ class Sun {
   /// See [meanAnomaly].
   double get uncorrectedMeanAnomaly =>
       _uncorrectedMeanAnomaly ??= evaluatePolynomial(
-          julianDate.julianCenturies, [357.52911, 35999.05029, -0.0001537]);
+          instant.julianCenturies, [357.52911, 35999.05029, -0.0001537]);
 
   /// The Sun's equation of center, in degrees.
   ///
@@ -111,9 +105,9 @@ class Sun {
   /// https://en.wikipedia.org/wiki/Equation_of_the_center
   double get equationOfCenter {
     if (_equationOfCenter == null) {
-      var mRadians = meanAnomaly.toRadians();
+      final mRadians = meanAnomaly.toRadians();
 
-      var julianCenturies = julianDate.julianCenturies;
+      final julianCenturies = instant.julianCenturies;
 
       _equationOfCenter = (sin(mRadians) *
               (1.914602 -
@@ -152,7 +146,7 @@ class Sun {
 
     if (_radiusVector == null) {
       // U.S. Naval Observatory function for radius vector
-      var mRadians = meanAnomaly.toRadians();
+      final mRadians = meanAnomaly.toRadians();
       _radiusVector =
           1.00014 - (0.01671 * cos(mRadians)) - (0.00014 * cos(2 * mRadians));
     }
@@ -166,7 +160,7 @@ class Sun {
   double get apparentLongitude => _apparentLongitude ??= trueLongitude -
       0.00569 -
       (0.00478 *
-          sin(Earth(julianDate).nutationAndAberrationCorrection.toRadians()));
+          sin(Earth(instant).nutationAndAberrationCorrection.toRadians()));
 
   /// The Apparent position of the Sun in the Equatorial Coordinate System.
   ///
@@ -176,15 +170,15 @@ class Sun {
   /// for the winter.
   EquatorialCoordinate get equatorialPosition {
     if (_equatorialPosition == null) {
-      var earth = Earth(julianDate);
+      final earth = Earth(instant);
 
-      var rightAscension = atan2(
+      final rightAscension = atan2(
         cos(earth.correctedObliquityOfEcliptic.toRadians()) *
             sin(apparentLongitude.toRadians()),
         cos(apparentLongitude.toRadians()),
       );
 
-      var declination = asin(
+      final declination = asin(
           sin(earth.correctedObliquityOfEcliptic.toRadians()) *
               sin(apparentLongitude.toRadians()));
 
@@ -197,12 +191,53 @@ class Sun {
 
   /// Gets the apparent position of the Sun in the Horizontal Coordinate System.
   HorizontalCoordinate horizontalPosition(double latitude, double longitude) {
-    var timeOffset = equationOfTime + (4.0 * longitude);
+    final hourAngle = _calculateLocalHourAngle(longitude);
+    final hourAngleRadians = hourAngle.toRadians();
 
-    var utcDateTime = julianDate.gregorianDateTime.toUtc();
-    var time = utcDateTime.time;
+    final latitudeRadians = latitude.toRadians();
 
-    var trueSolarTime = time.totalMinutes + timeOffset;
+    final sunDeclinationRadians = equatorialPosition.declination.toRadians();
+
+    final zenithRadians = _calculateSolarZenithAngle(
+        latitudeRadians, sunDeclinationRadians, hourAngleRadians);
+    final zenith = zenithRadians.toDegrees();
+
+    final azimuth = _calculateAzimuth(
+        latitudeRadians, zenithRadians, sunDeclinationRadians, hourAngle);
+    final elevation = _calculateElevation(zenith);
+
+    return HorizontalCoordinate(azimuth, elevation);
+  }
+
+  /// Calculates the solar zenith angle in radians.
+  ///
+  /// The [latitude], [declination] and [hourAngle] must be in radians.
+  double _calculateSolarZenithAngle(
+      double latitude, double declination, double hourAngle) {
+    final cosinusSolarZenith = (sin(latitude) * sin(declination)) +
+        (cos(latitude) * cos(declination) * cos(hourAngle));
+
+    if (cosinusSolarZenith > 1.0) return acos(1.0);
+    if (cosinusSolarZenith < -1.0) return acos(-1.0);
+
+    return acos(cosinusSolarZenith);
+  }
+
+  /// Calculates the Sun local hour angle in degrees for the given [longitude].
+  ///
+  /// The local hour angle is calculated from the local meridian.
+  ///
+  /// The [longitude] is given in degrees.
+  double _calculateLocalHourAngle(double longitude) {
+    final timeOffset = equationOfTime +
+        (4.0 * longitude) -
+        (Duration.minutesPerHour * instant.timeZoneOffset);
+
+    // final utcDateTime = instant.toUtc();
+    // final time = utcDateTime.time;
+
+    // var trueSolarTime = time.totalMinutes + timeOffset;
+    var trueSolarTime = instant.time.totalMinutes + timeOffset;
 
     while (trueSolarTime > Duration.minutesPerDay) {
       trueSolarTime -= Duration.minutesPerDay;
@@ -212,34 +247,39 @@ class Sun {
         (trueSolarTime / 4.0) - 180.0; // Solar hour angle in degrees
     if (hourAngle < -180) hourAngle += 360.0;
 
-    var latitudeRadians = latitude.toRadians();
-    var sunDeclinationRadians = equatorialPosition.declination.toRadians();
-    var hourAngleRadians = hourAngle.toRadians();
+    return hourAngle;
+  }
 
-    // Calculate the cosinus of the solar zenith angle.
-    var cosinusSolarZenith =
-        (sin(latitudeRadians) * sin(sunDeclinationRadians)) +
-            (cos(latitudeRadians) *
-                cos(sunDeclinationRadians) *
-                cos(hourAngleRadians));
+  /// Calculates the elevation of the Sun in degrees.
+  ///
+  /// [zenithAngle] is given in degrees.
+  double _calculateElevation(double zenithAngle) {
+    // Atmospheric Refraction correction
+    final refractionCorrection =
+        _calculateAtmosphericRefractionCorrection(90.0 - zenithAngle);
 
-    if (cosinusSolarZenith > 1.0) {
-      cosinusSolarZenith = 1.0;
-    } else if (cosinusSolarZenith < -1.0) cosinusSolarZenith = -1.0;
+    final correctedSolarZenith = zenithAngle - refractionCorrection;
+    return 90.0 - correctedSolarZenith;
+  }
 
-    var zenithRadians = acos(cosinusSolarZenith);
-    var zenith = zenithRadians.toDegrees();
-    var azimuthDenominator = cos(latitudeRadians) * sin(zenithRadians);
-
+  /// Calculates the azimuth of the Sun in degrees.
+  ///
+  /// The [latitude], [zenithAngle] and [declination] must be in radians.
+  /// The [hourAgngle] must be in degrees.
+  double _calculateAzimuth(double latitude, double zenithAngle,
+      double declination, double hourAngle) {
     double azimuth;
 
-    if (azimuthDenominator.abs() > 0.001) {
-      var azimuthRadians = ((sin(latitudeRadians) * cos(zenithRadians)) -
-              sin(sunDeclinationRadians)) /
-          azimuthDenominator;
+    final azimuthDenominator = cos(latitude) * sin(zenithAngle);
 
-      if (azimuthRadians.abs() > 1.0)
+    if (azimuthDenominator.abs() > 0.001) {
+      var azimuthRadians =
+          ((sin(latitude) * cos(zenithAngle)) - sin(declination)) /
+              azimuthDenominator;
+
+      if (azimuthRadians.abs() > 1.0) {
         azimuthRadians = (azimuthRadians < 0) ? -1.0 : 1.0;
+      }
 
       azimuth = 180.0 - acos(azimuthRadians).toDegrees();
 
@@ -250,14 +290,7 @@ class Sun {
 
     if (azimuth < 0.0) azimuth += 360.0;
 
-    // Atmospheric Refraction correction
-    var refractionCorrection =
-        _calculateAtmosphericRefractionCorrection(90.0 - zenith);
-
-    var correctedSolarZenith = zenith - refractionCorrection;
-    var elevation = 90.0 - correctedSolarZenith;
-
-    return HorizontalCoordinate(azimuth, elevation);
+    return azimuth;
   }
 
   /// Calculates the approximate Atmospheric Refraction Correction in degrees for the given [elevation].
@@ -268,23 +301,21 @@ class Sun {
   ///
   /// https://www.esrl.noaa.gov/gmd/grad/solcalc/calcdetails.html
   double _calculateAtmosphericRefractionCorrection(double elevation) {
-    if (elevation > 85.0) {
-      return 0.0;
+    if (elevation > 85.0) return 0.0;
+
+    final tanElevation = tan(elevation.toRadians());
+    double correction;
+
+    if (elevation > 5.0) {
+      correction = evaluatePolynomial(
+          1 / tanElevation, [0, 58.1, 0, -0.07, 0, 0.000086]);
+    } else if (elevation > -0.575) {
+      correction =
+          evaluatePolynomial(elevation, [1735, -518.2, 103.4, -12.79, 0.711]);
     } else {
-      var tanElevation = tan(elevation.toRadians());
-      double correction;
-
-      if (elevation > 5.0) {
-        correction = evaluatePolynomial(
-            1 / tanElevation, [0, 58.1, 0, -0.07, 0, 0.000086]);
-      } else if (elevation > -0.575) {
-        correction =
-            evaluatePolynomial(elevation, [1735, -518.2, 103.4, -12.79, 0.711]);
-      } else {
-        correction = -20.774 / tanElevation;
-      }
-
-      return correction / 3600.0;
+      correction = -20.774 / tanElevation;
     }
+
+    return correction / 3600.0;
   }
 }